Rename cdc module to etl
