--- cdc/src/main/java/com/shzhangji/vault/cdc/HikariDataSourceFactory.java
@@ -1,10 +1,0 @@
-package com.shzhangji.vault.cdc;
-
-import com.zaxxer.hikari.HikariDataSource;
-import org.apache.ibatis.datasource.unpooled.UnpooledDataSourceFactory;
-
-public class HikariDataSourceFactory extends UnpooledDataSourceFactory {
-  public HikariDataSourceFactory() {
-    this.dataSource = new HikariDataSource();
-  }
-}
