--- cdc/src/main/java/com/shzhangji/vault/cdc/extractconfig/DbInstanceMapper.java
@@ -1,7 +1,0 @@
-package com.shzhangji.vault.cdc.extractconfig;
-
-import java.util.Optional;
-
-public interface DbInstanceMapper {
-  Optional<DbInstanceRow> get(int id);
-}
