--- cdc/src/main/java/com/shzhangji/vault/cdc/extractconfig/ExtractConfigMapper.java
@@ -1,7 +1,0 @@
-package com.shzhangji.vault.cdc.extractconfig;
-
-import java.util.List;
-
-public interface ExtractConfigMapper {
-  List<ExtractConfigRow> getList(int tenantId);
-}
