--- cdc/src/main/java/com/shzhangji/vault/cdc/extractconfig/TenantId.java
@@ -1,5 +1,0 @@
-package com.shzhangji.vault.cdc.extractconfig;
-
-public class TenantId {
-  public static final int VAULT = 1;
-}
