--- cdc/src/main/java/com/shzhangji/vault/cdc/sourcetable/PartitionExtractor.java
@@ -1,8 +1,0 @@
-package com.shzhangji.vault.cdc.sourcetable;
-
-import com.shzhangji.vault.cdc.sourcetable.SourceTable;
-import java.util.Map;
-
-public interface PartitionExtractor {
-  Map<String, String> extract(SourceTable sourceTable, Map<String, String> columns);
-}
