--- cdc/src/main/java/com/shzhangji/vault/cdc/sourcetable/RowKind.java
@@ -1,6 +1,0 @@
-package com.shzhangji.vault.cdc.sourcetable;
-
-public enum RowKind {
-  REPLACE,
-  DELETE,
-}
