--- cdc/src/main/java/com/shzhangji/vault/cdc/sourcetable/SourceTable.java
@@ -1,10 +1,0 @@
-package com.shzhangji.vault.cdc.sourcetable;
-
-import lombok.Value;
-
-@Value
-public class SourceTable {
-  int instanceId;
-  String database;
-  String table;
-}
