--- cdc/src/main/java/com/shzhangji/vault/cdc/targettable/TargetTable.java
@@ -1,9 +1,0 @@
-package com.shzhangji.vault.cdc.targettable;
-
-import lombok.Value;
-
-@Value
-public class TargetTable {
-  String database;
-  String table;
-}
