--- cdc/src/main/resources/mybatis-config.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE configuration
-  PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-  "http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-  <settings>
-    <setting name="mapUnderscoreToCamelCase" value="true"/>
-  </settings>
-  <environments default="vault">
-    <environment id="vault">
-      <transactionManager type="JDBC"/>
-      <dataSource type="com.shzhangji.vault.cdc.HikariDataSourceFactory">
-        <property name="jdbcUrl" value="${jdbc.url}"/>
-        <property name="username" value="${jdbc.username}"/>
-        <property name="password" value="${jdbc.password}"/>
-        <property name="poolName" value="vault"/>
-        <property name="minimumIdle" value="0"/>
-      </dataSource>
-    </environment>
-  </environments>
-  <mappers>
-    <mapper resource="com/shzhangji/vault/cdc/extractconfig/DbInstanceMapper.xml"/>
-    <mapper resource="com/shzhangji/vault/cdc/extractconfig/ExtractConfigMapper.xml"/>
-  </mappers>
-</configuration>
