+++ etl/pom.xml
@@ -11,11 +11,37 @@
   <modelVersion>4.0.0</modelVersion>
   <artifactId>vault-etl</artifactId>
 
+  <properties>
+    <flink.version>1.17.1</flink.version>
+  </properties>
+
   <dependencies>
     <dependency>
       <groupId>com.shzhangji.vault</groupId>
       <artifactId>vault-common</artifactId>
       <version>0.0.1-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.google.inject</groupId>
+      <artifactId>guice</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.mybatis</groupId>
+      <artifactId>mybatis</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.zaxxer</groupId>
+      <artifactId>HikariCP</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.flink</groupId>
+      <artifactId>flink-streaming-java</artifactId>
+      <version>${flink.version}</version>
+      <scope>provided</scope>
     </dependency>
   </dependencies>
 </project>
