+++ etl/src/main/java/com/shzhangji/vault/etl/HikariDataSourceFactory.java
@@ -1,0 +1,10 @@
+package com.shzhangji.vault.etl;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.apache.ibatis.datasource.unpooled.UnpooledDataSourceFactory;
+
+public class HikariDataSourceFactory extends UnpooledDataSourceFactory {
+  public HikariDataSourceFactory() {
+    this.dataSource = new HikariDataSource();
+  }
+}
