+++ etl/src/main/java/com/shzhangji/vault/etl/extractconfig/DbInstanceMapper.java
@@ -1,0 +1,7 @@
+package com.shzhangji.vault.etl.extractconfig;
+
+import java.util.Optional;
+
+public interface DbInstanceMapper {
+  Optional<DbInstanceRow> get(int id);
+}
