+++ etl/src/main/java/com/shzhangji/vault/etl/extractconfig/ExtractConfigMapper.java
@@ -1,0 +1,7 @@
+package com.shzhangji.vault.etl.extractconfig;
+
+import java.util.List;
+
+public interface ExtractConfigMapper {
+  List<ExtractConfigRow> getList(int tenantId);
+}
