+++ etl/src/main/java/com/shzhangji/vault/etl/extractconfig/ExtractConfigRow.java
@@ -1,0 +1,22 @@
+package com.shzhangji.vault.etl.extractconfig;
+
+import java.util.Date;
+import lombok.Data;
+
+@Data
+public class ExtractConfigRow {
+  private int sourceInstanceId;
+  private String sourceDatabase;
+  private String sourceTable;
+  private boolean isRegex;
+  private String primaryKeys;
+  private String extractColumns;
+  private int targetInstanceId;
+  private String targetDatabase;
+  private String targetTable;
+  private int partitionType;
+  private String partitionInputColumn;
+  private String partitionOutputColumn;
+  private boolean ignoreDelete;
+  private Date updatedAt;
+}
