+++ etl/src/main/java/com/shzhangji/vault/etl/extractconfig/TenantId.java
@@ -1,0 +1,5 @@
+package com.shzhangji.vault.etl.extractconfig;
+
+public class TenantId {
+  public static final int VAULT = 1;
+}
