+++ etl/src/main/java/com/shzhangji/vault/etl/sourcetable/PartitionExtractor.java
@@ -1,0 +1,7 @@
+package com.shzhangji.vault.etl.sourcetable;
+
+import java.util.Map;
+
+public interface PartitionExtractor {
+  Map<String, String> extract(SourceTable sourceTable, Map<String, String> columns);
+}
