+++ etl/src/main/java/com/shzhangji/vault/etl/sourcetable/RowKind.java
@@ -1,0 +1,6 @@
+package com.shzhangji.vault.etl.sourcetable;
+
+public enum RowKind {
+  REPLACE,
+  DELETE,
+}
