+++ etl/src/main/java/com/shzhangji/vault/etl/sourcetable/SourceTable.java
@@ -1,0 +1,10 @@
+package com.shzhangji.vault.etl.sourcetable;
+
+import lombok.Value;
+
+@Value
+public class SourceTable {
+  int instanceId;
+  String database;
+  String table;
+}
