+++ etl/src/main/java/com/shzhangji/vault/etl/targettable/TargetTable.java
@@ -1,0 +1,9 @@
+package com.shzhangji.vault.etl.targettable;
+
+import lombok.Value;
+
+@Value
+public class TargetTable {
+  String database;
+  String table;
+}
