+++ etl/src/main/resources/com/shzhangji/vault/etl/extractconfig/DbInstanceMapper.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.shzhangji.vault.etl.extractconfig.DbInstanceMapper">
+  <select id="get" resultType="com.shzhangji.vault.etl.extractconfig.DbInstanceRow">
+    SELECT
+      `id`
+      ,`host`
+      ,`port`
+      ,`username`
+      ,`password`
+      ,`database`
+    FROM db_instance
+    WHERE id = #{id}
+  </select>
+</mapper>
