+++ etl/src/main/resources/com/shzhangji/vault/etl/extractconfig/ExtractConfigMapper.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.shzhangji.vault.etl.extractconfig.ExtractConfigMapper">
+  <select id="getList" resultType="com.shzhangji.vault.etl.extractconfig.ExtractConfigRow">
+    SELECT
+      source_instance_id
+      ,source_database
+      ,source_table
+      ,is_regex
+      ,primary_keys
+      ,extract_columns
+      ,target_instance_id
+      ,target_database
+      ,target_table
+      ,partition_type
+      ,partition_input_column
+      ,partition_output_column
+      ,ignore_delete
+      ,updated_at
+    FROM extract_config
+    WHERE status = 1
+    AND tenant_id = #{tenantId}
+    ORDER BY id
+  </select>
+</mapper>
